backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.10.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0
msal==1.31.0
httpx==0.27.2
pydantic==2.9.2
python-multipart==0.0.12
